requirements.txt
requests
lxml
beautifulsoup4
whitenoise
google-cloud-logging
web.py==0.40
gunicorn